requirements.txt
discord.py
google-generativeai
python-dotenv # Optional, for local testing
